fix:useSearchParams Suspense, add EmailStep and PasswordStep components to handle login flow
